Increase timeout in http_model_client.py
Some users are reporting timeouts with 70b models, we're capping running times by saying the total runtime of docker image needs to be 2h but people are free to take longer on individual examples if they so choose

```
We are trying to use llama 70b, but occasionally encounter the following request problems during the reasoning process: Request error: HTTPConnectionPool(host='localhost', port=8080): Read timed out. (read timeout=300), Can the single request time be extended? Or do we have to complete it in 300 seconds? Thank you for your help
```