Move get_run_instances behavior from Adapter to runner.py
Currently all adapters share this behavior, and the only use of `self` is getting max_eval_instances. IMO no adapter should have different behavior in downsampling eval instances, so removing that flexibility is a good idea. Finally, this cleans up the division of responsibilities, as Adapter gets to be a pure interface.