Calculate num_instances in its own Metric. (#2191)
This dramatically simplifies the complexity of what should be a simple Metric. For example, previously:

1. The value reported is per-train trial, but only because:
    1. get_request_states reports an empty list for instances not in the train_trial
    2. that then causes those instances to have 0 Stats
    3. When adding to grouped_per_instance_stats we use a for loop on `append` so it doesn't get an entry for that instance.
4. If you somehow produced zero Stats for an Instance a different way, that Instance wouldn't be counted in num_instances.

To check end-to-end if this is changing outputs, I ran`mmlu:subject=philosophy,model=openai/gpt2` with `--max-eval-instances 1000` at `main` and at this branch. The only diff in `stats.json` was where in the list `num_instances` appears.